+++ saffier/contrib/admin/__init__.py
@@ -1,5 +1,7 @@
 from .application import create_admin_app
 from .config import AdminConfig
+from .controllers import admin_not_found
+from .mixins import AdminMixin
 from .site import AdminSite
 
-__all__ = ["AdminConfig", "AdminSite", "create_admin_app"]
+__all__ = ["AdminConfig", "AdminMixin", "AdminSite", "admin_not_found", "create_admin_app"]