+++ saffier/contrib/lilya/__init__.py
@@ -1,0 +1,3 @@
+from .middleware import EdgyMiddleware
+
+__all__ = ["EdgyMiddleware"]