+++ saffier/core/db/querysets/__init__.py
@@ -1,5 +1,5 @@
-from .base import QuerySet
+from .base import CombinedQuerySet, QuerySet
 from .clauses import Q, and_, not_, or_
 from .prefetch import Prefetch
 
-__all__ = ["QuerySet", "Prefetch", "Q", "and_", "not_", "or_"]
+__all__ = ["CombinedQuerySet", "QuerySet", "Prefetch", "Q", "and_", "not_", "or_"]