+++ saffier/core/utils/__init__.py
@@ -1,3 +1,20 @@
 from .concurrency import batched, run_concurrently
+from .db import (
+    CHECK_DB_CONNECTION_SILENCED,
+    FORCE_FIELDS_NULLABLE,
+    check_db_connection,
+    force_fields_nullable_as_list_string,
+    hash_names,
+    hash_tablekey,
+)
 
-__all__ = ["batched", "run_concurrently"]
+__all__ = [
+    "CHECK_DB_CONNECTION_SILENCED",
+    "FORCE_FIELDS_NULLABLE",
+    "batched",
+    "check_db_connection",
+    "force_fields_nullable_as_list_string",
+    "hash_names",
+    "hash_tablekey",
+    "run_concurrently",
+]